+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
