+++ src/test/java/roomescape/controller/ReservationPageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.Test;
