+++ src/test/java/roomescape/dao/DBConnectionTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
